+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/service/LinkService.java
@@ -5,8 +5,8 @@ import ru.tinkoff.edu.java.scrapper.repository.LinkRecord;
 import java.util.List;
 
 public interface LinkService {
-    void add(String url, int chatId);
-    void delete(String url, int chatId);
-    List<LinkRecord> linksForChat(int chatId);
+    void add(String url, long chatId);
+    void delete(String url, long chatId);
+    List<LinkRecord> linksForChat(long chatId);
     List<LinkRecord> getAll();
 }
